+++ src/components/profile/analysisCard/chart/Chart.Style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Colors, FontStyles } from '../../../styles';
+import { Colors, FontStyles } from '../../../../styles';
 
 export const ChartContainer = styled.div`
   width: 100%;
