+++ src/components/profile/analysisCard/chart/Chart.tsx
