+++ src/components/profile/analysisCard/grade/Grade.tsx
