+++ src/components/profile/profileCard/coinExplain/CoinExplain.Style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Colors, FontStyles } from '../../../styles';
+import { Colors, FontStyles } from '../../../../styles';
 
 export const Explain = styled.div`
   position: absolute;
